Extract TemplateCache to templates/_cache.dart
